v0.1.5.2 20241113 删除了暂时未使用的代码 在主程序添加chip参数，修改README 整理主程序格式，决定后端实现形式

+++ ir/dialect/top/IR_tensor.py
@@ -77,7 +77,7 @@ class IRTensor:  # IR中，表示张量的class
     Name = "top_ir_tensor"
     Id = None
     Type = TensorType.Intermediate
-    DataType = DataType.INT8
+    DataType = DataType.FLOAT32
     ZeroPoint = None
     Scale = None
     Q_min = None
